add more tests for different combinations of path properties
